api/requirements_api.txt
# Web/API
fastapi>=0.101.0
uvicorn[standard]>=0.23.1
python-multipart>=0.0.6

# Data & ML
numpy>=1.24.0

# Model Interoperability
onnxruntime-cpu

# Hugging Face
huggingface_hub>=0.17.0
